refactor(utils): rename utils to epoch for clarity
