+++ src/lib.rs
@@ -3,4 +3,4 @@ pub mod errors;
 pub mod inclusion_proof;
 pub mod sync;
 pub mod types;
-pub mod utils;
+pub mod epoch;
